refactor: rename pool fees split percent for incoming schema updates

+++ contract/script/pools-manager/DeployCastoraPoolsManager.s.sol
@@ -12,7 +12,7 @@ contract DeployCastoraPoolsManager is Script {
     address castoraAddress = 0xa0742C672e713327b0D6A4BfF34bBb4cbb319C53; // vm.envAddress('CASTORA_ADDRESS');
     address poolsRulesAddress = 0x00Ae1567c73f74b8445A013FD71E518E2EFD326e; // vm.envAddress('POOLS_RULES_ADDRESS');
     address feeCollectorAddress = vm.envAddress('FEE_COLLECTOR_ADDRESS');
-    uint256 splitPercent = 5000; // 50% , 2 decimals // vm.envUint('COMPLETION_FEES_SPLIT_PERCENT');
+    uint16 splitPercent = 5000; // 50% , 2 decimals // vm.envUint('COMPLETION_FEES_SPLIT_PERCENT');
 
     address proxy = Upgrades.deployUUPSProxy(
       'CastoraPoolsManager.sol',
